moved headers to inside the pages
