+++ package.json
@@ -1,10 +1,9 @@
 {
   "name": "prototype",
   "version": "1.0.0",
-  "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "npx tailwindcss -i ./src/input.css -o ./src/output.css --watch"
+    "start": "npx tailwindcss -i ./src/utils/input.css -o ./src/utils/output.css --watch"
   },
   "keywords": [],
   "author": "",
