+++ src/profile/profile.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Touringo</title>
+    <title>Touringo Profile Page</title>
     <script src="https://cdn.tailwindcss.com"></script>
     <link rel="stylesheet" href="../utils/output.css">
 </head> 
@@ -46,18 +46,24 @@
             </div>     
         </div>
     </div>
-
-    <div id="homeID" class="dark:bg-gradient-to-b dark:from-gray-800 p-4 bg-gradient-to-t h-max flex flex-col">
-        <div class="max-w-max mx-auto relative">
-            <div class="py-4">
-                <h1 class="text-3xl border-b-[20px] border-blue-300 pb-4">
-                    See attractions below
-                </h1>
-                <table id="app" class="dark:bg-blue-900 bg-blue-100 p-2 table-auto text-center border-separate border-spacing-2 border border-slate-500"></table>                
-            </div>
+    <!-- Profile container -->
+    <div class="profile-container max-w-3xl mx-auto mt-1 my-8 bg-gray-200 p-6 rounded-lg shadow-lg text-center">
+        <div class="profile-header mb-6">
+            <img id="profile-pic" class="rounded-full w-36 h-36 object-cover border-4 border-blue-500 mx-auto" src="../../assets/images/profilePicture.jpg" alt="Profile Picture">
+            <h1 id="name" class="mt-4 text-2xl text-blue-500">John Doe</h1>
+            <p id="bio" class="text-lg text-gray-600">Web Developer | Tech Enthusiast | Lifelong Learner</p>
         </div>
+        <div class="profile-content mb-6">
+            <h2 class="text-xl text-blue-500">About Me</h2>
+            <p id="about" class="mt-2 text-gray-700">
+                I am a passionate web developer with experience in building responsive and dynamic web applications. I love exploring new technologies and sharing knowledge with others.
+            </p>
+        </div>
+        <button id="edit-btn" class="mt-6 px-6 py-2 bg-blue-500 text-white rounded-lg hover:bg-blue-700">
+            Edit Profile
+        </button>
     </div>
 
-    <script type="module" src="./home.js"></script>
+    <script type="module" src="./profile.js"></script>
 </body>
 </html>
