--- src/userProfile/profile.html
@@ -1,42 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Profile Page</title>
-    <link rel="stylesheet" href="profile.css">
-    <script src="profile.js"></script>
-</head>
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Profile with Iframe</title>
-    <script src="https://cdn.tailwindcss.com"></script>
-</head>
-<body class="bg-gray-700 h-screen ">
-    <!-- Iframe container, takes up remaining space -->
-    <div class="flex-grow w-full">
-        <iframe class="w-full border-0" src="../headers/headers.html"></iframe>
-    </div>
-
-    <!-- Profile container -->
-    <div class="profile-container max-w-3xl mx-auto mt-1 my-8 bg-gray-200 p-6 rounded-lg shadow-lg text-center">
-        <div class="profile-header mb-6">
-            <img id="profile-pic" class="rounded-full w-36 h-36 object-cover border-4 border-blue-500 mx-auto" src="../../assets/images/profilePicture.jpg" alt="Profile Picture">
-            <h1 id="name" class="mt-4 text-2xl text-blue-500">John Doe</h1>
-            <p id="bio" class="text-lg text-gray-600">Web Developer | Tech Enthusiast | Lifelong Learner</p>
-        </div>
-        <div class="profile-content mb-6">
-            <h2 class="text-xl text-blue-500">About Me</h2>
-            <p id="about" class="mt-2 text-gray-700">
-                I am a passionate web developer with experience in building responsive and dynamic web applications. I love exploring new technologies and sharing knowledge with others.
-            </p>
-        </div>
-        <button id="edit-btn" class="mt-6 px-6 py-2 bg-blue-500 text-white rounded-lg hover:bg-blue-700">
-            Edit Profile
-        </button>
-    </div>
-</body>
-</html>
